states and transitions now use pointer to struct State as value
Using a string that is the name of the state was good but an much more
idiomatic solution is to get access to the state object through the
tranisition member of a state.

The previous attempt used the State object as the value which resulted
in infinite recursion when a state had a transition function to
itself.

Now the code has been rewritten to use pointers to State objects as
the value of states and transitions. Since the hashtables now just
store the address instead of the entire object, infinite recursion
does not occur.